refactor: extract shared helpers into generate_utils module
Consolidates five duplicated helper functions (ancestor_of_kind,
leading_whitespace, extract_primary_ctor_params, existing_method_names,
find_insert_position, make_action) from all three generate-* feature
files into a single generate_utils module.

Also fixes:
- Double iteration over definition_locations in collect_overridable_methods
- Missing blank-line separators between methods in combined actions
- Unused import of KIND_CLASS_DECL in generate_utils
